:new: [확장] 메뉴 관리 기능 추가 #28  Changes to be committed: 	modified:   src/main/java/com/eun/tutorial/config/MyWebSecurityConfigurerAdapter.java 	new file:   src/main/java/com/eun/tutorial/controller/main/MenuController.java 	new file:   src/main/java/com/eun/tutorial/dto/main/MenuDTO.java 	modified:   src/main/java/com/eun/tutorial/filter/XssFilter.java 	new file:   src/main/java/com/eun/tutorial/mapper/main/MenuMapper.java 	new file:   src/main/java/com/eun/tutorial/service/main/MenuService.java 	new file:   src/main/java/com/eun/tutorial/service/main/MenuServiceImpl.java 	modified:   src/main/java/com/eun/tutorial/util/OracleTypeConverter.java 	new file:   src/main/resources/mybatis/mapper/main/MenuMapper.xml 	modified:   src/main/resources/schema.sql 	modified:   src/main/resources/static/js/main/common/main.js 	new file:   src/main/resources/static/js/main/content/menu.js 	new file:   src/main/webapp/WEB-INF/jsp/main/common/leftNav2.jsp 	modified:   src/main/webapp/WEB-INF/jsp/main/common/main.jsp 	new file:   src/main/webapp/WEB-INF/jsp/main/content/menu.jsp

+++ src/main/java/com/eun/tutorial/util/OracleTypeConverter.java
@@ -12,7 +12,7 @@ public class OracleTypeConverter {
             oracleType.equalsIgnoreCase("CLOB")) {
             return String.class;
         } else if (oracleType.equalsIgnoreCase("NUMBER")) {
-            return BigDecimal.class; // 또는 int, long, double 등 숫자 타입 중 하나로 매핑 가능
+            return int.class; // 또는 int, long, double 등 숫자 타입 중 하나로 매핑 가능
         } else if (oracleType.equalsIgnoreCase("DATE") ||
                    oracleType.equalsIgnoreCase("TIMESTAMP")) {
             return Timestamp.class; // 또는 java.util.Date, java.time.LocalDateTime 등으로 매핑 가능
